Created Login requirement and welcome page for a successful Login. Website does not allow access to homepage or any editing features without a successful Login.

+++ app/app.py
@@ -1,11 +1,13 @@
 from typing import List, Dict
 import simplejson as json
 from flask import Flask, request, Response, redirect
-from flask import render_template
+from flask import render_template, g, session, url_for, flash
 from flaskext.mysql import MySQL
 from pymysql.cursors import DictCursor
 
+
 app = Flask(__name__)
+app.secret_key = 'brandonsecretkey'
 mysql = MySQL(cursorclass=DictCursor)
 
 app.config['MYSQL_DATABASE_HOST'] = 'db'
@@ -16,8 +18,64 @@ app.config['MYSQL_DATABASE_DB'] = 'deniroData'
 mysql.init_app(app)
 
 
-@app.route('/', methods=['GET'])
+class User:
+    def __init__(self, id, username, password):
+        self.id = id
+        self.username = username
+        self.password = password
+
+    def __repr__(self):
+        return f'<User: {self.username}>'
+
+
+users = [User(id=1, username='Class', password='password')]
+
+
+@app.before_request
+def before_request():
+    g.user = None
+
+    if 'user_id' in session:
+        user = [x for x in users if x.id == session['user_id']][0]
+        g.user = user
+
+
+@app.route('/', methods=['GET', 'POST'])
+def login():
+    if request.method == 'POST':
+        session.pop('user_id', None)
+
+        username = request.form['username']
+        password = request.form['password']
+
+        userlist = [x for x in users if x.username == username]
+        if len(userlist) > 0:
+            user = userlist[0]
+        else:
+            return redirect(url_for('login'))
+
+        if user and user.password == password:
+            session['user_id'] = user.id
+            flash("Login Successful")
+            return redirect(url_for('profile'))
+
+        return redirect(url_for('login'))
+    flash("Welcome! Please Login to Proceed!")
+    return render_template('login.html')
+
+
+@app.route('/profile')
+def profile():
+    if not g.user:
+        return redirect(url_for('login'))
+
+    return render_template('profile.html')
+
+
+@app.route('/index', methods=['GET'])
 def index():
+    if not g.user:
+        return redirect(url_for('login'))
     user = {'username': 'Deniro Movies Project'}
     cursor = mysql.get_db().cursor()
     cursor.execute('SELECT * FROM tblDeniroImport')
@@ -27,6 +85,8 @@ def index():
 
 @app.route('/view/<int:movie_id>', methods=['GET'])
 def record_view(movie_id):
+    if not g.user:
+        return redirect(url_for('login'))
     cursor = mysql.get_db().cursor()
     cursor.execute('SELECT * FROM tblDeniroImport WHERE id=%s', movie_id)
     result = cursor.fetchall()
@@ -35,6 +95,8 @@ def record_view(movie_id):
 
 @app.route('/edit/<int:movie_id>', methods=['GET'])
 def form_edit_get(movie_id):
+    if not g.user:
+        return redirect(url_for('login'))
     cursor = mysql.get_db().cursor()
     cursor.execute('SELECT * FROM tblDeniroImport WHERE id=%s', movie_id)
     result = cursor.fetchall()
@@ -43,6 +105,8 @@ def form_edit_get(movie_id):
 
 @app.route('/edit/<int:movie_id>', methods=['POST'])
 def form_update_post(movie_id):
+    if not g.user:
+        return redirect(url_for('login'))
     cursor = mysql.get_db().cursor()
     inputData = (request.form.get('Title'), request.form.get('Year'), request.form.get('Score'),
                  movie_id)
@@ -51,13 +115,18 @@ def form_update_post(movie_id):
     mysql.get_db().commit()
     return redirect("/", code=302)
 
+
 @app.route('/movies/new', methods=['GET'])
 def form_insert_get():
+    if not g.user:
+        return redirect(url_for('login'))
     return render_template('new.html', title='New Movie Form')
 
 
 @app.route('/movies/new', methods=['POST'])
 def form_insert_post():
+    if not g.user:
+        return redirect(url_for('login'))
     cursor = mysql.get_db().cursor()
     inputData = (request.form.get('Title'), request.form.get('Year'), request.form.get('Score'))
     sql_insert_query = """INSERT INTO tblDeniroImport (Title,Year,Score) VALUES (%s, %s,%s) """
@@ -65,8 +134,11 @@ def form_insert_post():
     mysql.get_db().commit()
     return redirect("/", code=302)
 
+
 @app.route('/delete/<int:movie_id>', methods=['POST'])
 def form_delete_post(movie_id):
+    if not g.user:
+        return redirect(url_for('login'))
     cursor = mysql.get_db().cursor()
     sql_delete_query = """DELETE FROM tblDeniroImport WHERE id = %s """
     cursor.execute(sql_delete_query, movie_id)
